feat: Updated README.md
Added the project background information, installation instructions, and
a disclaimer.

Obviously the nature of a program that finds the locations of people's
IP addresses will draw the attention of malevolent actors seeking to use
CLIp for illit purposes. This can't really be prevented by the tool's
very nature, so the least I can do is to have a disclaimer discouraging
using CLIp for such purposes.